+++ services/account/service/src/main/java/org/collectionspace/services/account/storage/csidp/UserStorageClient.java
@@ -64,7 +64,7 @@ public class UserStorageClient {
         User user = new User();
         user.setUsername(userId);
         String salt = UUID.randomUUID().toString();
-        user.setPasswd(getEncPassword(userId, password, salt));
+        user.setPasswd(getEncPassword(password));
         user.setSalt(salt);
         user.setCreatedAt(new Date());
         return user;
@@ -116,7 +116,7 @@ public class UserStorageClient {
         User userFound = get(jpaTransactionContext, userId);
         if (userFound != null) {
             String salt = UUID.randomUUID().toString();
-            userFound.setPasswd(getEncPassword(userId, password, salt));
+            userFound.setPasswd(getEncPassword(password));
             userFound.setSalt(salt);
             userFound.setUpdatedAt(new Date());
             if (logger.isDebugEnabled()) {
@@ -150,15 +150,7 @@ public class UserStorageClient {
         }
     }
 
-    private String getEncPassword(String userId, byte[] password, String salt) throws BadRequestException {
-        //jaxb unmarshaller already unmarshal xs:base64Binary, no need to b64 decode
-        //byte[] bpass = Base64.decodeBase64(accountReceived.getPassword());
-        try {
-            SecurityUtils.validatePassword(new String(password));
-        } catch (Exception e) {
-            throw new BadRequestException(e.getMessage());
-        }
-        String secEncPasswd = SecurityUtils.createPasswordHash(new String(password));
-        return secEncPasswd;
+    private String getEncPassword(byte[] password) {
+        return SecurityUtils.createPasswordHash(new String(password));
     }
 }